refactor: improve type safety and code clarity in cursor and coordinate handling
- Updated Makefile to enhance compiler optimization flags for better performance.
- Cast coordinate calculations to 'short' type in deck and field modules to ensure type safety and prevent potential overflow issues.
- Refactored map initialization in main.c to use explicit struct initialization for clarity.

These changes enhance code readability and maintainability, ensuring safer coordinate handling and improved performance in the game engine.

+++ impl/solitaire/src/deck.c
@@ -62,7 +62,7 @@ static void place_cursor_in_deck(const void *deck_pointer, const Coords cursor_c
     (void)deck_pointer;
     
     target_coords->y = BORDER_OFFSET_Y + CARD_HEIGHT;
-    target_coords->x = cursor_coords.x * CARD_WIDTH + (CARD_WIDTH / 2) + BORDER_OFFSET_X - 1;
+    target_coords->x = (short)(cursor_coords.x * CARD_WIDTH + (CARD_WIDTH / 2) + BORDER_OFFSET_X - 1);
 }
 
 /*
